Fix Package/UObject compile error

+++ SurrealEngine/Editor/Export.cpp
@@ -340,7 +340,7 @@ MemoryStreamWriter Exporter::ExportBmpIndexed(UTexture* tex)
 		data << bgra;
 	}
 
-	hdr.pixelOffset = data.Tell();
+	hdr.pixelOffset = (uint32_t)data.Tell();
 
 	uint8_t *pixels = tex->Mipmaps[0].Data.data();
 	for (int y = vsize; y > 0; y--)
